Refactor BaGet (#261)
1. Moves business logic out of controllers and into `BaGet.Core`. For example, the `PackageContentController` passes requests to `IBaGetPackageContentService` and maps results to HTTP responses. In effect, controllers are just translate HTTP translators.
2. BaGet's services for the NuGet protocol implement the same interfaces as BaGet's client libraries for the NuGet protocol. For example, `AzureSearchService` and `SearchClient` both implement `ISearchService`
3. Requests and responses are codified by models in `BaGet.Protocol`. For example, both `SearchController.SearchAsync` and `SearchClient.SearchAsync` return a `SearchResponse`

+++ src/BaGet.Protocol/Search/SearchResponse.cs
@@ -11,7 +11,7 @@ namespace BaGet.Protocol
     public class SearchResponse
     {
         public SearchResponse(
-            int totalHits,
+            long totalHits,
             IReadOnlyList<SearchResult> data,
             SearchContext context = null)
         {
@@ -26,7 +26,7 @@ namespace BaGet.Protocol
         /// <summary>
         /// The total number of matches, disregarding skip and take.
         /// </summary>
-        public int TotalHits { get; }
+        public long TotalHits { get; }
 
         /// <summary>
         /// The packages that matched the search query.
